+++ main/managers/views/options_screen.c
@@ -126,7 +126,7 @@ static const char *gps_options[] = {"Start Wardriving", "Stop Wardriving", "GPS 
                                     "BLE Wardriving",   "Go Back",         NULL};
 
 static const char *settings_options[] = {"Set RGB Mode - Stealth", "Set RGB Mode - Normal",
-                                         "Set RGB Mode - Rainbow", "Go Back", NULL};
+                                         "Set RGB Mode - Rainbow", "Webauth On", "Webauth Off", "Go Back", NULL};
 
 
 static void up_down_event_cb(lv_event_t *e) {
@@ -733,6 +733,18 @@ void option_event_cb(lv_event_t *e) {
         simulateCommand("setsetting 1 3");
         vTaskDelay(pdMS_TO_TICKS(10));
         error_popup_create("Set RGB Mode Successfully...");
+        
+    }
+
+    else if (strcmp(Selected_Option, "Webauth On") == 0) {
+        simulateCommand("webauth on");
+        error_popup_create("Webauth Enabled.");
+        
+    }
+
+    else if (strcmp(Selected_Option, "Webauth Off") == 0) {
+        simulateCommand("webauth off");
+        error_popup_create("Webauth Disabled.");
         
     }