+++ CHANGELOG.md
@@ -31,6 +31,7 @@
 - Commands
   - Allow selection of multiple APs (eg. select -a 2,3,4)
   TODO - add comma to display numpad
+  - webauthcommand to enable/disable web authentication
 
 - Cardputer
   - Add keyboard event handling functionality - @tototo31